+++ CHANGELOG.md
@@ -37,3 +37,4 @@
 - [feat:add admin http handler.](https://github.com/Tencent/spring-cloud-tencent/pull/1449)
 - [feat:support concurrency rate limit.](https://github.com/Tencent/spring-cloud-tencent/pull/1456)
 - [feat:support auth.](https://github.com/Tencent/spring-cloud-tencent/pull/1460)
+- [feat:upgrade trace plugin.](https://github.com/Tencent/spring-cloud-tencent/pull/1468)